Moved BW adaptation to adapt_schemes, added configurable scripts
