+++ erizo_controller/erizoJS/erizoJSController.js
@@ -21,8 +21,6 @@ exports.ErizoJSController = function (spec) {
 
         INTERVAL_TIME_SDP = 100,
         INTERVAL_TIME_KILL = 30*60*1000, // Timeout to kill itself after a timeout since the publisher leaves room.
-        INTERVAL_STATS = 1000,
-        MIN_RECOVER_BW = 50000,
         SLIDESHOW_TIME = 1000,
         initWebRtcConnection,
         getSdp,
@@ -33,146 +31,7 @@ exports.ErizoJSController = function (spec) {
 
     var CONN_INITIAL = 101, CONN_STARTED = 102,CONN_GATHERED = 103, CONN_READY = 104, CONN_FINISHED = 105, CONN_CANDIDATE = 201, CONN_SDP = 202, CONN_FAILED = 500;
 
-    /* BW Status
-     * 0 - Stable 
-     * 1 - Insufficient Bandwidth 
-     * 2 - Trying recovery
-     * 3 - Won't recover
-     */
 
-    var BW_STABLE = 0, BW_INSUFFICIENT = 1, BW_RECOVERING = 2, BW_WONTRECOVER = 3;
-
-    var calculateAverage = function (values) { 
-        if (values.length === undefined)
-            return 0;
-        var cnt = values.length;
-        var tot = parseInt(0);
-        for (var i = 0; i < values.length; i++){
-            tot+=parseInt(values[i]);
-        }
-        return Math.ceil(tot/cnt);
-    };
-
-    /*
-     * Monitors BW from a subscriber and reacts accordingly
-     */
-
-    monitorMinVideoBw = function(wrtc, callback){
-        wrtc.bwValues = [];
-        var onlyNotifyBW= true; //TODO: For now we only notify, study if we want to enable the rest
-        var isReporting = true;
-        var ticks = 0;
-        var retries = 0;
-        var ticksToTry = 0;
-        var lastAverage, average, lastBWValue, toRecover;
-        var nextRetry = 0;
-        wrtc.bwStatus = BW_STABLE;
-        log.debug("START Monitoring Video BW", wrtc.minVideoBW);
-
-        wrtc.minVideoBW = wrtc.minVideoBW*1000; // We need it in bps
-        wrtc.lowerThres = Math.floor(wrtc.minVideoBW*(1-0.2));
-        wrtc.upperThres = Math.ceil(wrtc.minVideoBW);
-        var intervalId = setInterval(function () {
-            var newStats = wrtc.getStats();
-            if (newStats == null){
-                log.debug("Stopping stats");
-                clearInterval(intervalId);
-                return;
-            }
-
-            var theStats = JSON.parse(newStats);
-            for (var i = 0; i < theStats.length; i++){ 
-                if(theStats[i].hasOwnProperty('bandwidth')){   // Only one stream should have bandwidth
-                    lastBWValue = theStats[i].bandwidth;
-                    wrtc.bwValues.push(lastBWValue);
-                    if (wrtc.bwValues.length > 5){
-                        wrtc.bwValues.shift();
-                    }
-                    average = calculateAverage(wrtc.bwValues);
-                }
-            }
-            toRecover = (average/4)<MIN_RECOVER_BW?(average/4):MIN_RECOVER_BW;
-            switch (wrtc.bwStatus){
-                case BW_STABLE:
-                    if(average <= lastAverage && (average < wrtc.lowerThres)){
-                        if (++ticks > 2){
-                            log.debug("STABLE STATE, Bandwidth is insufficient, moving to state BW_INSUFFICIENT", average, "lowerThres", wrtc.lowerThres);
-                            if (!onlyNotifyBW){
-                                wrtc.bwStatus = BW_INSUFFICIENT;
-                                wrtc.setFeedbackReports(false, toRecover);
-                            }
-                            ticks = 0;
-                            callback('callback', {type:'bandwidthAlert', message:'insufficient', bandwidth: average});
-                        }
-                    }                            
-                    break;
-                case BW_INSUFFICIENT:
-                    if(average > wrtc.upperThres){
-                        log.debug("BW_INSUFFICIENT State: we have recovered", average, "lowerThres", wrtc.lowerThres);
-                        ticks = 0;
-                        nextRetry = 0;
-                        retries = 0;
-                        wrtc.bwStatus = BW_STABLE;
-                        wrtc.setFeedbackReports(true, 0);
-                        callback('callback', {type:'bandwidthAlert', message:'recovered', bandwidth: average});
-                    }
-                    else if (retries>=3){
-                        log.debug("BW_INSUFFICIENT State: moving to won't recover", average, "lowerThres", wrtc.lowerThres);
-                        wrtc.bwStatus = BW_WONTRECOVER; 
-                    }
-                    else if (nextRetry === 0){  //schedule next retry
-                        nextRetry = ticks + 20;
-                    }
-                    else if (++ticks == nextRetry){  // next retry is in order
-                        wrtc.bwStatus = BW_RECOVERING;
-                        ticksToTry = ticks + 10;
-                        wrtc.setFeedbackReports (false, average);                                
-                    }
-                    break;
-                case BW_RECOVERING:
-                    log.debug("In recovering state lastValue", lastBWValue, "lastAverage", lastAverage, "lowerThres", wrtc.lowerThres);
-                    if(average > wrtc.upperThres){ 
-                        log.debug("BW_RECOVERING State: we have recovered", average, "lowerThres", wrtc.lowerThres);
-                        ticks = 0;
-                        nextRetry = 0;
-                        retries = 0;
-                        wrtc.bwStatus = BW_STABLE;
-                        wrtc.setFeedbackReports(true, 0);
-                        callback('callback', {type:'bandwidthAlert', message:'recovered', bandwidth: average});
-                    }
-                    else if (average> lastAverage){ //we are recovering
-                        log.debug("BW_RECOVERING State: we have improved, more trying time", average, "lowerThres", wrtc.lowerThres);
-                        wrtc.setFeedbackReports(false, average*(1+0.3));
-                        ticksToTry=ticks+10;
-
-                    }
-                    else if (++ticks >= ticksToTry){ //finish this retry
-                        log.debug("BW_RECOVERING State: Finished this retry", retries, average, "lowerThres", wrtc.lowerThres);
-                        ticksToTry = 0;
-                        nextRetry = 0;
-                        retries++;
-                        wrtc.bwStatus = BW_INSUFFICIENT;
-                        wrtc.setFeedbackReports (false, toRecover);
-                    }
-                    break;
-                case BW_WONTRECOVER:
-                    log.debug("BW_WONTRECOVER", average, "lowerThres", wrtc.lowerThres);
-                    ticks = 0;
-                    nextRetry = 0;
-                    retries = 0;
-                    average = 0;
-                    lastAverage = 0;
-                    wrtc.bwStatus = BW_STABLE;
-                    wrtc.minVideoBW = false;                      
-                    wrtc.setFeedbackReports (false, 1);
-                    callback('callback', {type:'bandwidthAlert', message:'wont-recover', bandwidth: average});
-                    break;
-                default:
-                    log.error("Unknown BW status");
-            }
-            lastAverage = average;
-        }, INTERVAL_STATS);
-    };
 
 
     /*
@@ -218,7 +77,19 @@ exports.ErizoJSController = function (spec) {
      * Given a WebRtcConnection waits for the state CANDIDATES_GATHERED for set remote SDP.
      */
     initWebRtcConnection = function (wrtc, callback, id_pub, id_sub, options) {
+        log.debug("Initing WebRTC Connection, minVideoBW", wrtc.minVideoBW);
         if (wrtc.minVideoBW){
+            var monitorMinVideoBw = {};
+            if (wrtc.scheme){
+                try{
+                    monitorMinVideoBw = require("./adapt_schemes/"+options.scheme).MonitorSubscriber;
+                } catch (e){
+                    log.warn("Could not find adapt_scheme script, will use notify");
+                    monitorMinVideoBw = require("./adapt_schemes/notify").MonitorSubscriber(log);
+                }
+            }else{
+                monitorMinVideoBw = require("./adapt_schemes/notify").MonitorSubscriber(log);
+            }
             monitorMinVideoBw(wrtc, callback);
         }
 
@@ -425,13 +296,13 @@ exports.ErizoJSController = function (spec) {
 
         if (publishers[from] === undefined) {
 
-            log.info("Adding publisher peer_id ", from, "minVideoBW", options.minVideoBW, "createOffer", options.createOffer);
+            log.info("Adding publisher peer_id ", from, "minVideoBW", options.minVideoBW,"adaptation scheme", options.scheme);
 
             var muxer = new addon.OneToManyProcessor(),
                 wrtc = new addon.WebRtcConnection(true, true, GLOBAL.config.erizo.stunserver, GLOBAL.config.erizo.stunport, GLOBAL.config.erizo.minport, GLOBAL.config.erizo.maxport,false,
                         GLOBAL.config.erizo.turnserver, GLOBAL.config.erizo.turnport, GLOBAL.config.erizo.turnusername, GLOBAL.config.erizo.turnpass);
 
-            publishers[from] = {muxer: muxer, wrtc: wrtc, minVideoBW: options.minVideoBW};
+            publishers[from] = {muxer: muxer, wrtc: wrtc, minVideoBW: options.minVideoBW, scheme:options.scheme};
             subscribers[from] = {};
 
             wrtc.setAudioReceiver(muxer);
@@ -483,6 +354,7 @@ exports.ErizoJSController = function (spec) {
         subscribers[to][from] = wrtc;
         publishers[to].muxer.addSubscriber(wrtc, from);
         wrtc.minVideoBW = publishers[to].minVideoBW;
+        wrtc.scheme = publishers[to].scheme;
         initWebRtcConnection(wrtc, callback, to, from, options);
     };
